# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechAtlas API
  description: ''
  contact:
    name: TechAtlas Contributors
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
paths:
...
  /api/v1/admin/domains:
    get:
      tags:
      - admin-domains
      operationId: list_domains
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Page of active domains
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainListResponse'
        '401':
          description: Authentication failed
          headers:
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer authentication challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Query validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
    post:
      tags:
      - admin-domains
      operationId: create_domain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainRequest'
        required: true
      responses:
        '201':
          description: Domain created
          headers:
            Location:
              schema:
                type: string
              description: Canonical resource location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainResponse'
        '401':
          description: Authentication failed
          headers:
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer authentication challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
  /api/v1/admin/domains/{canonical_domain}:
    get:
      tags:
      - admin-domains
      operationId: get_domain
      parameters:
      - name: canonical_domain
        in: path
        description: Domain name or normalizable HTTP(S) URL
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Active domain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainResponse'
        '401':
          description: Authentication failed
          headers:
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer authentication challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Path validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
...
components:
  schemas:
...
    CreateDomainRequest:
      type: object
      required:
      - domain
      properties:
        domain:
          type: string
...
    DomainListResponse:
      type: object
      required:
      - domains
      properties:
        domains:
          type: array
          items:
            $ref: '#/components/schemas/DomainResponse'
        next_cursor:
          type:
          - string
          - 'null'
    DomainResponse:
      type: object
      required:
      - id
      - canonical_domain
      properties:
        canonical_domain:
          type: string
        id:
          type: string
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
  securitySchemes:
    oidc_bearer:
      type: http
      scheme: bearer
      bearerFormat: Bearer